style(auth): move validators to separate files
